+++ src/raster/wasm/terrano_wasm_bg.wasm.d.ts
@@ -13,6 +13,7 @@ export const polygonize: (a: number, b: number, c: number, d: number, e: number,
 export const rasterize: (a: number, b: number, c: number, d: number, e: number, f: number, g: number, h: number) => [number, number, number, number];
 export const reclassify: (a: number, b: number, c: number, d: number, e: number, f: number, g: number, h: number) => [number, number, number, number];
 export const slope: (a: number, b: number, c: number, d: number, e: number, f: number) => [number, number, number, number];
+export const writeCog: (a: number, b: number, c: number, d: number, e: number, f: number, g: number, h: number, i: number, j: number, k: number, l: number, m: number, n: number, o: number) => [number, number, number, number];
 export const zonalStats: (a: number, b: number, c: number, d: number, e: number, f: number, g: number, h: number) => [number, number, number, number];
 export const __wbindgen_externrefs: WebAssembly.Table;
 export const __wbindgen_malloc: (a: number, b: number) => number;